WIN32: Move getCodePageId to codepage.h
