+++ backends/platform/sdl/module.mk
@@ -23,8 +23,7 @@ MODULE_OBJS += \
 	win32/win32-main.o \
 	win32/win32-window.o \
 	win32/win32_wrapper.o \
-	win32/win32.o \
-	win32/codepage.o
+	win32/win32.o
 endif
 
 ifdef AMIGAOS
