--- backends/platform/sdl/win32/codepage.cpp
@@ -1,39 +1,0 @@
-/* ScummVM - Graphic Adventure Engine
- *
- * ScummVM is the legal property of its developers, whose names
- * are too numerous to list here. Please refer to the COPYRIGHT
- * file distributed with this source distribution.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-
-#ifdef WIN32
-#include "backends/platform/sdl/win32/codepage.h"
-namespace Win32 {
-
-int getCodePageId(Common::String codePageName) {
-	const CodePageDescription *cp = g_cpDescriptions;
-	for (; cp->name; cp++) {
-		if (codePageName.equalsIgnoreCase(cp->name))
-			return cp->id;
-	}
-	return -1;
-}
-
-}
-
-#endif
-
